requirements.txt
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
click==7.1.2
cryptography==2.8
decorator==4.4.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.3.2
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.3
psycopg2==2.8.5
py-trello==0.15.0
pycparser==2.19
pystache==0.5.4
python-dateutil==2.8.1
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.16
tornado==6.0.3
urllib3==1.25.6
virtualenv==16.4.3
Werkzeug==1.0.1
